feat: refactor and handle providers in the right way
Refactor to a more functional paradigm for all of the functionality.
Avoiding `this` should make it so that we can avoid all of the possible
errors that come with this.

Also, the code should be more readable now.

The big piece of this feature is making choosing the right provider 
work.
Previously, if you had more than one "provider" then it would not
properly go to definition.